requirements.txt
Flask>=2.0
Werkzeug
mysql-connector-python
